+++ packages/@eris-boiler/common/src/index.ts
@@ -1,11 +1,10 @@
 // modules
 export * from '@modules/loadable'
-export * from '@modules/extended-map'
-export * as logger from '@modules/logger'
+export * from '@modules/message'
+export * from '@modules/permissions'
 
 // utils
-export * from '@utils/constructor'
+export * from '@utils/extended-map'
 export * from '@utils/identity'
-export * from '@utils/keys-matching'
-export * from '@utils/message'
-export * from '@utils/permissions'
+export * as logger from '@utils/logger'
+export * from '@utils/type'
