--- packages/@eris-boiler/common/src/utils/constructor.ts
@@ -1,1 +1,0 @@
-export type Constructor<T> = new (...args: any[]) => T
