--- packages/@eris-boiler/common/src/utils/keys-matching.ts
@@ -1,3 +1,0 @@
-export type KeysMatching<T, V> = {
-  [K in keyof T]-?: T[K] extends V ? K : never
-}[keyof T]
